[feat] (WIP) Attempt to write all commits in a .txt file
----------
- ISSUES -
----------

- Github's API request rate is limited for non-fully authentificated requests (What a fricking bummer),
so now I have to input my API token and make a system to extract secrets from keybase or something etc...
That will for sure delay any progress on the rest of the project for at least half a day at my estimation.

- I can't verify if my code ACTUALLY works or not because of the issue mentionned above. So we'll have the
incertitude until I fix this.

- My code to change the API's page is a bit junky, but I think it'll work for now. Though I will still need
to improve this. It'll become an issue when when have tens of thousands commits in a repository.